Fixed synchronization bug in download properties dialog.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@9065 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/BTTrackerRequest.h
@@ -264,7 +264,7 @@ public:
 	CBTTrackerRequest* Lookup(DWORD nTransactionID) const;
 
 protected:
-	typedef CMap< WORD, WORD, CBTTrackerRequest*, CBTTrackerRequest* > CBTTrackerRequestMap;
+	typedef CMap< DWORD, DWORD, CBTTrackerRequest*, CBTTrackerRequest* > CBTTrackerRequestMap;
 
 	mutable CCriticalSection	m_pSection;
 	CBTTrackerRequestMap		m_pTrackerRequests;	// Tracker ID to tracker pointer map
